backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
PyJWT==2.8.0

python-dotenv==1.0.0
werkzeug==3.0.1
requests==2.31.0
google-generativeai==0.8.5
webauthn==2.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
